Move yaptu code to own module
Refactoring to improve code maintainability.
