+++ gcinvoice.py
@@ -29,6 +29,7 @@ import logging
 import optparse
 import ConfigParser
 from operator import itemgetter
+from yaptu import copier
 get0 = itemgetter(0)
 get_entered = itemgetter('entered')
 
@@ -475,8 +476,8 @@ class Gcinvoice(object):
             # The name of the template file is itself a template in order to
             # select different templates depending on the invoice.
             templ_ = StringIO.StringIO()
-            cop = _copier(rex, invc, rbe, ren, rco, ouf=templ_,
-                          encoding=self._gcfile_encoding)
+            cop = copier(rex, invc, rbe, ren, rco, ouf=templ_,
+                         encoding=self._gcfile_encoding)
             cop.copy([template])
             templ = templ_.getvalue()
             templ_.close()
@@ -520,8 +521,8 @@ class Gcinvoice(object):
             # The name of the outfile is itself a template in order to
             # select different outfiles depending on the invoice.
             outf_ = StringIO.StringIO()
-            cop = _copier(rex, invc, rbe, ren, rco, ouf=outf_,
-                          encoding=self._gcfile_encoding)
+            cop = copier(rex, invc, rbe, ren, rco, ouf=outf_,
+                         encoding=self._gcfile_encoding)
             cop.copy([outfile])
             outfile = outf_.getvalue()
             outf_.close()
@@ -544,8 +545,8 @@ class Gcinvoice(object):
             self.logger.warn("Cannot do template for expression [%s]" % expr,
                              exc_info=True)
             return expr
-        cop = _copier(rex, invc, rbe, ren, rco, ouf=outf, handle=handle,
-                      encoding=self._gcfile_encoding)
+        cop = copier(rex, invc, rbe, ren, rco, ouf=outf, handle=handle,
+                     encoding=self._gcfile_encoding)
         try:
             cop.copy(templ)
         except Exception:
@@ -861,108 +862,7 @@ class _MyTemplate(Template):
     idpattern = '[_a-z][-_a-z0-9]*'
 
 
-# This is the Yet Another Python Templating Utility, Version 1.2
-# Taken from the ActiveState python Cookbook recipe 52305
-# by Alex Martelli
-# Adapted by Roman Bertle for the needs of this module.
-
-# utility stuff to avoid tests in the mainline code
-class _nevermatch:
-    "Polymorphic with a regex that never matches"
-    def match(self, line):
-        return None
-
-
-_never = _nevermatch()     # one reusable instance of it suffices
-
-
-def _identity(string, why):
-    "A do-nothing-special-to-the-input, just-return-it function"
-    return string
-
-
-def _nohandle(string):
-    "A do-nothing handler that just re-raises the exception"
-    raise
-
-
-# and now the real thing
-class _copier:
-    "Smart-copier (YAPTU) class"
-    def copyblock(self, i=0, last=None):
-        "Main copy method: process lines [i,last) of block"
-        def repl(match, self=self):
-            "return the eval of a found expression, for replacement"
-            # uncomment for debug: print '!!! replacing',match.group(1)
-            expr = self.preproc(match.group(1), 'eval')
-            try:
-                return unicode(eval(expr, self.globals, self.locals))
-            except:
-                return unicode(self.handle(expr))
-        block = self.locals['_bl']
-        if last is None:
-            last = len(block)
-        while i < last:
-            line = block[i]
-            match = self.restat.match(line)
-            if match:   # a statement starts "here" (at line block[i])
-                # i is the last line to _not_ process
-                stat = match.string[match.end(0):].strip()
-                j = i+1   # look for 'finish' from here onwards
-                nest = 1  # count nesting levels of statements
-                while j < last:
-                    line = block[j]
-                    # first look for nested statements or 'finish' lines
-                    if self.restend.match(line):    # found a statement-end
-                        nest = nest - 1     # update (decrease) nesting
-                        if nest == 0:
-                            break   # j is first line to _not_ process
-                    elif self.restat.match(line):   # found a nested statement
-                        nest = nest + 1     # update (increase) nesting
-                    elif nest == 1:
-                        # look for continuation only at this nesting
-                        match = self.recont.match(line)
-                        if match:                   # found a contin.-statement
-                            nestat = match.string[match.end(0):].strip()
-                            stat = '%s _cb(%s,%s)\n%s' % (stat, i+1, j, nestat)
-                            i = j  # again, i is the last line to _not_ process
-                    j = j+1
-                stat = self.preproc(stat, 'exec')
-                stat = '%s _cb(%s,%s)' % (stat, i+1, j)
-                # for debugging, uncomment...: print "-> Executing: {"+stat+"}"
-                exec(stat, self.globals, self.locals)
-                i = j+1
-            else:       # normal line, just copy with substitution
-                self.ouf.write(self.regex.sub(repl, line).encode(
-                    self.encoding))
-                i = i+1
-
-    def __init__(self, regex=_never, dict={},
-                 restat=_never, restend=_never, recont=_never,
-                 preproc=_identity, handle=_nohandle, ouf=sys.stdout,
-                 encoding='ascii'):
-        "Initialize self's attributes"
-        self.regex = regex
-        self.globals = dict
-        self.locals = {'_cb': self.copyblock}
-        self.restat = restat
-        self.restend = restend
-        self.recont = recont
-        self.preproc = preproc
-        self.handle = handle
-        self.ouf = ouf
-        self.encoding = encoding
-
-    def copy(self, block=None, inf=sys.stdin):
-        "Entry point: copy-with-processing a file, or a block of lines"
-        if block is None:
-            block = inf.readlines()
-        self.locals['_bl'] = block
-        self.copyblock()
-
-
 # Things if the module is run as script
-
 
 def createInvoice(invoiceid, template=None, outfile=None, options=None):
